adds faust streamer to pipeline: events to pr-events
